Optimize and condense
Sorry, I'm afraid I don't speak your language.

The `bash` builtin `printf` is more than capable of such date and time
fetching, and setting the format you wish for the backup filename. The
`date` command is best left for more advanced usages, or when using the
Bourne Shell (`sh`).

The `find` command itself can execute commands. By using `+` at the
end of the `-exec` part, it's being told to append each path to the
same command, rather than executing a new process with each file.

Rather than using `gzip` and `tar` separately, they can be combined by
using the `-z` flag within `tar`. There are other forms of compression
of which tar can make use, so be sure to check its man page for more
information.

The `exit 0` at the end of the script is redundant, as this is exactly
what would happen regardless, unless a command failed, in which case
you probably _want_ to see the non-zero exit status (for diagnosis).

Used `printf` for the message to the user, for portability due to its
consistent behavior.